+++ app/src/main/java/com/example/android_repo_04/view/listener/RefreshListener.kt
@@ -1,0 +1,5 @@
+package com.example.android_repo_04.view.listener
+
+interface RefreshListener {
+    fun showProgress()
+}
