--- app/src/main/java/com/example/android_repo_04/view/search/SearchRefreshListener.kt
@@ -1,5 +1,0 @@
-package com.example.android_repo_04.view.search
-
-interface SearchRefreshListener {
-    fun showProgress()
-}
